requirements.txt
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cffi==2.0.0
click==8.3.1
cssselect2==0.9.0
dnspython==2.8.0
email-validator==2.3.0
eventlet==0.40.4
Flask==3.1.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.62.1
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
SQLAlchemy==2.0.48
tinycss2==1.5.1
tinyhtml5==2.1.0
typing_extensions==4.15.0
weasyprint==68.1
webencodings==0.5.1
Werkzeug==3.1.6
wsproto==1.3.2
WTForms==3.2.1
zopfli==0.4.1
